fix bug theme

+++ app/src/main/java/com/example/selenaapp/ui/settings/SettingsFragment.kt
@@ -8,6 +8,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.CompoundButton
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.datastore.core.DataStore
@@ -22,6 +23,7 @@ import com.example.selenaapp.data.preference.dataStore
 import com.example.selenaapp.databinding.FragmentSettingsBinding
 import com.example.selenaapp.ui.help.HelpActivity
 import com.example.selenaapp.ui.login.LoginActivity
+import com.example.selenaapp.ui.main.MainActivity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
@@ -45,17 +47,16 @@ class SettingsFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         (activity as? AppCompatActivity)?.supportActionBar?.hide()
 
-        val switchTheme = binding.switch2
-        val pref = SettingsPreference.getInstance(requireActivity().dataStore)
-        val settingViewModel = ViewModelProvider(this, SettingsViewModelFactory(pref))
-            .get(SettingsViewModel::class.java)
+        val pref = SettingsPreference.getInstance(requireContext().dataStore)
+        val viewModelFactory = SettingsViewModelFactory(pref)
+        val settingViewModel = ViewModelProvider(this, viewModelFactory).get(SettingsViewModel::class.java)
 
         // Observe changes to theme setting
         settingViewModel.getThemeSettings().observe(viewLifecycleOwner) { isDarkModeActive ->
-            switchTheme.isChecked = isDarkModeActive // Update the UI based on current setting
+            binding.switch2.isChecked = isDarkModeActive
         }
 
-        switchTheme.setOnCheckedChangeListener { _, isChecked ->
+        binding.switch2.setOnCheckedChangeListener { _, isChecked ->
             settingViewModel.saveThemeSetting(isChecked) // Save theme setting, don't apply directly here
         }
 
@@ -75,13 +76,25 @@ class SettingsFragment : Fragment() {
     private fun setupListeners() {
         // Tombol logout
         binding.btnLogout.setOnClickListener {
-            viewLifecycleOwner.lifecycleScope.launch {
-                val userPreference = UserPreference.getInstance(requireContext().dataStore)
-                userPreference.logout()
-                val intent = Intent(requireContext(), LoginActivity::class.java)
-                intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-                startActivity(intent)
+            AlertDialog.Builder(requireContext()).apply {
+            setTitle("Hapus Semua Transaksi")
+            setMessage("Anda yakin ingin menghapus semua transaksi?")
+            setPositiveButton("Ya") { _, _ ->
+                viewLifecycleOwner.lifecycleScope.launch {
+                    val userPreference = UserPreference.getInstance(requireContext().dataStore)
+                    userPreference.logout()
+                    val intent = Intent(requireContext(), LoginActivity::class.java)
+                    intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+                    startActivity(intent)
+                }
             }
+            setNegativeButton("Tidak") { _, _ ->
+                //do nothing
+            }
+            create()
+            show()
+        }
+
         }
 
         // Tombol help
@@ -92,7 +105,18 @@ class SettingsFragment : Fragment() {
 
         // Tombol hapus semua transaksi
         binding.btnDeleteAll.setOnClickListener {
-            deleteAllTransaction()
+            AlertDialog.Builder(requireContext()).apply {
+                setTitle("Keluar")
+                setMessage("Anda yakin ingin Logout dari aplikasi?")
+                setPositiveButton("Ya") { _, _ ->
+                    deleteAllTransaction()
+                }
+                setNegativeButton("Tidak") { _, _ ->
+                    //do nothing
+                }
+                create()
+                show()
+            }
         }
     }
 
@@ -109,6 +133,9 @@ class SettingsFragment : Fragment() {
                 withContext(Dispatchers.Main) {
                     if (response.isSuccessful) {
                         showToast(response.body()?.message ?: "Berhasil menghapus semua transaksi.")
+                        val intent = Intent(context, MainActivity::class.java)
+                        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+                        startActivity(intent)
                     } else {
                         showToast(response.body()?.message ?: "Gagal menghapus transaksi.")
                     }
@@ -132,6 +159,8 @@ class SettingsFragment : Fragment() {
         _binding = null
     }
 }
+
+
